Modulus of negative dividends is undefined or negative
... depending on who you ask. Since it is possible for applicationPid to
return negative values this means we would introduce garbage ['()*+,-./]
in the generated filenames.

(cherry picked from commit cb7cb1d3884ae8a032f3ad2ed3a6d8e3ffc06206)

+++ src/corelib/io/qtemporaryfile.cpp
@@ -118,7 +118,7 @@ static int createFileFromTemplate(char *const path,
         char *rIter = placeholderEnd;
 
 #if defined(QT_BUILD_CORE_LIB)
-        qint64 pid = QCoreApplication::applicationPid();
+        quint64 pid = quint64(QCoreApplication::applicationPid());
         do {
             *--rIter = (pid % 10) + '0';
             pid /= 10;
